refactor(cli)!: restructure CLI into subcommand-first architecture with serve/db commands
Reorganize phoenix.server.main into a thin arg-parser dispatcher that
delegates to phoenix.server.cli.commands.{serve,db}.  This enables
standalone `db migrate` for running migrations without starting the
server — useful in container init scripts and CI pipelines.

BREAKING CHANGE: CLI flags now follow the subcommand instead of
preceding it.  For example:
  Before: phoenix --dev serve
  After:  phoenix serve --dev

Other changes:
- Remove deprecated --enable-websockets flag
- Move welcome banner into app lifespan so it prints after startup
- Update dev scripts (package.json, start-phoenix.sh) for new arg order
- Relocate grpc port tests to tests/unit/server/cli/commands/test_serve.py
- Document migration steps in MIGRATION.md (v14)
