+++ app/package.json
@@ -13,10 +13,10 @@
     "dev": "pnpm run ensure:env && mprocs",
     "dev:embeddings": "pnpm run dev:server:embeddings & pnpm run build:static && pnpm run build:relay && source ./.env && vite",
     "dev:offline": "pnpm run dev:server:offline & pnpm run build:static && pnpm run build:relay && source ./.env && vite",
-    "dev:server": "source ./.env && uv run python -Xfrozen_modules=off -m debugpy --listen ${DEBUGPY_PORT:-5678} -m phoenix.server.main --dev --dev-vite-port ${VITE_PORT:-5173} serve",
+    "dev:server": "source ./.env && uv run python -Xfrozen_modules=off -m debugpy --listen ${DEBUGPY_PORT:-5678} -m phoenix.server.main serve --dev --dev-vite-port ${VITE_PORT:-5173}",
     "dev:server:embeddings": "pnpm dev:server --with-fixture=fashion_mnist",
     "dev:server:init": "pnpm dev:server --with-trace-fixtures=llama_index_rag --with-projects=SESSIONS-DEMO --force-fixture-ingestion",
-    "dev:server:offline": "python -m phoenix.server.main --dev --dev-vite-port ${VITE_PORT:-5173} serve",
+    "dev:server:offline": "python -m phoenix.server.main serve --dev --dev-vite-port ${VITE_PORT:-5173}",
     "dev:server:test": "node ./tests/utils/testServer.mjs",
     "dev:ui": "source ./.env && pnpm run build:static && pnpm run build:relay && vite",
     "ensure:env": "test -f .env || printf 'export PHOENIX_TELEMETRY_ENABLED=False\nexport PHOENIX_DANGEROUSLY_ENABLE_AGENTS=True\n' > .env",
